+++ src/handlers/scheduled-handler.ts
@@ -4,6 +4,7 @@ import { directCastsHandler } from '@/handlers/direct-casts-handler'
 import { eventsHandler } from '@/handlers/events-handler'
 import { proposalHandler } from '@/handlers/proposal-handler'
 import { reminderHandler } from '@/handlers/reminder-handler'
+import { starterPackHandler } from '@/handlers/starter-pack-handler'
 import { logger } from '@/utilities/logger'
 import { CronTime } from 'cron-time-generator'
 
@@ -25,6 +26,7 @@ export async function scheduledHandler(
       await reminderHandler(env)
       break
     case CronTime.every(12).hours():
+      await starterPackHandler(env)
       await directCastsHandler(env)
       break
     case CronTime.everyDayAt(14, 0):